requirements.txt
flask
numpy
pandas
yfinance
scikit-learn
plotly
tensorflow-cpu
